+++ R/all_potential_biomarkers.R
@@ -9,11 +9,11 @@ all_potential_biomarkers <- function(dataset) {
   missingness_vars <- grep("missingness", names(dataset), value = TRUE)
   tremors <- c(
     "head tremor hx",
-    "Intention tremor",
-    "Resting tremor",
-    "Postural tremor",
-    "Intermittent tremor hx",
-    "Any tremor (excluding head)"
+    "intention tremor",
+    "resting tremor",
+    "postural tremor",
+    "intermittent tremor hx",
+    "any tremor (excluding head)"
   )
 
   parkinsonian_vars <-
@@ -21,7 +21,7 @@ all_potential_biomarkers <- function(dataset) {
       "parkinsonian features",
       "Masked faces",
       "Increased tone",
-      "Pill-rolling tremor",
+      "pill-rolling tremor",
       "Stiff gait"
     )
 
@@ -103,10 +103,10 @@ all_potential_biomarkers <- function(dataset) {
 
   kinesia_vars <- c(
     "Kinesia Left Rest Tremor*",
-    "Kinesia Left Postural Tremor*",
+    "Kinesia Left postural tremor*",
     "Kinesia Left Kinetic Tremor*",
     "Kinesia Right Rest Tremor*",
-    "Kinesia Right Postural Tremor*",
+    "Kinesia Right postural tremor*",
     "Kinesia Right Kinetic Tremor*"
   )