+++ R/create_any_tremor.R
@@ -1,18 +1,24 @@
 create_any_tremor <- function(
     dataset,
     tremor_types = c(
-      "intention tremor hx",
-      "resting tremor hx",
-      "postural tremor hx",
-      "Intermittent tremor hx"
+      "intention tremor",
+      "resting tremor",
+      "postural tremor",
+      "intermittent tremor"
     )
 )
 {
   dataset |>
     dplyr::mutate(
-      "Any tremor (excluding head)" = case_when(
-        dplyr::if_any(.cols = all_of(tremor_types), .fns = ~ . %in% "Yes") ~ "Some tremors recorded",
-        dplyr::if_all(.cols = all_of(tremor_types), .fns = ~ is.na(.) | . %in% c("NA (888)", "No Response (999)")) ~ NA,
+      "any tremor (excluding head)" = case_when(
+        dplyr::if_any(
+          .cols = all_of(tremor_types),
+          .fns = ~ . %in% "Yes"
+        ) ~ "Some tremors recorded",
+        dplyr::if_all(
+          .cols = all_of(tremor_types),
+          .fns = ~ is.na(.) | . %in% c("NA (888)", "No Response (999)")
+        ) ~ NA,
         TRUE ~ "No tremors recorded"
       ) |>
         factor() |>